refactor: extracts storage types into separate files
