+++ storage/record.go
@@ -1,0 +1,10 @@
+package storage
+
+import (
+	"time"
+)
+
+type Record struct {
+	Date    time.Time
+	Entries []Entry
+}
